Separate C BIOS handlers into bios.c from blinkenlights.c
